Turned select traveler into it's own component
